app/requirements.txt
Django==5.1.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
factory_boy==3.3.1
Faker==33.0.0
